GO-4408 Rename date util pkg
