+++ core/block/detailservice/relations.go
@@ -18,7 +18,7 @@ import (
 	coresb "github.com/anyproto/anytype-heart/pkg/lib/core/smartblock"
 	"github.com/anyproto/anytype-heart/pkg/lib/database"
 	"github.com/anyproto/anytype-heart/space/spacecore/typeprovider"
-	"github.com/anyproto/anytype-heart/util/date"
+	"github.com/anyproto/anytype-heart/util/dateutil"
 	"github.com/anyproto/anytype-heart/util/pbtypes"
 	"github.com/anyproto/anytype-heart/util/slice"
 )
@@ -114,7 +114,7 @@ func generateFilter(value *types.Value) func(v *types.Value) bool {
 		return equalFilter
 	}
 
-	start, err := date.ParseDateId(stringValue)
+	start, err := dateutil.ParseDateId(stringValue)
 	if err != nil {
 		log.Error("failed to convert date id to day start", zap.Error(err))
 		return equalFilter
