+++ core/block/detailservice/relations_test.go
@@ -16,7 +16,7 @@ import (
 	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
 	"github.com/anyproto/anytype-heart/pkg/lib/localstore/objectstore"
 	"github.com/anyproto/anytype-heart/pkg/lib/pb/model"
-	"github.com/anyproto/anytype-heart/util/date"
+	"github.com/anyproto/anytype-heart/util/dateutil"
 	"github.com/anyproto/anytype-heart/util/pbtypes"
 )
 
@@ -57,7 +57,7 @@ func TestService_ListRelationsWithValue(t *testing.T) {
 		{
 			bundle.RelationKeyId:               pbtypes.String("obj2"),
 			bundle.RelationKeySpaceId:          pbtypes.String(spaceId),
-			bundle.RelationKeyName:             pbtypes.String(date.TimeToDateId(now)),
+			bundle.RelationKeyName:             pbtypes.String(dateutil.TimeToDateId(now)),
 			bundle.RelationKeyCreatedDate:      pbtypes.Int64(now.Add(-24*time.Hour - 5*time.Minute).Unix()),
 			bundle.RelationKeyAddedDate:        pbtypes.Int64(now.Add(-24*time.Hour - 3*time.Minute).Unix()),
 			bundle.RelationKeyLastModifiedDate: pbtypes.Int64(now.Add(-1 * time.Minute).Unix()),
@@ -84,13 +84,13 @@ func TestService_ListRelationsWithValue(t *testing.T) {
 	}{
 		{
 			"date object - today",
-			pbtypes.String(date.TimeToDateId(now)),
+			pbtypes.String(dateutil.TimeToDateId(now)),
 			[]string{bundle.RelationKeyAddedDate.String(), bundle.RelationKeyCreatedDate.String(), bundle.RelationKeyLastModifiedDate.String(), bundle.RelationKeyName.String()},
 			[]int64{1, 2, 3, 1},
 		},
 		{
 			"date object - yesterday",
-			pbtypes.String(date.TimeToDateId(now.Add(-24 * time.Hour))),
+			pbtypes.String(dateutil.TimeToDateId(now.Add(-24 * time.Hour))),
 			[]string{bundle.RelationKeyAddedDate.String(), bundle.RelationKeyCreatedDate.String()},
 			[]int64{1, 1},
 		},
